Fixes #151
If you have a multi-segment path defined (/api/v1/proxy/{path*}) it breaks when it creates the validator due to it creating the param name as "path*" instead of "path". As multi-segment parameters can also have a repeat variable (path*2 as example) replacing the path variable name needs to take this into account.