Upgrade hapi to v19
This upgrade required two main changes:

1. There is no longer a way to query the strategies registered with the
server. The previous behavior was to warn that a strategy was not
registered, then continue creating the route without the specified auth
options. With this change, the route options will be added as long as
they are specified. If the auth strategy is not registered, then an
error will be thrown when registering the route. This seems to be inline
with README.

2. Hapi 19 no longer allows uncompiled top-level validation schemas
unless a validator is configured with the server. In previous versions,
Joi was used by default. To address this, after the validators are
created for a path/operation, any top-level object that is not already a
Joi schema is wrapped in Joi.object().