Upgrade dependencies to support hapi 19 (#173)
* Upgrade hapi to v19

This upgrade required two main changes:

1. There is no longer a way to query the strategies registered with the
server. The previous behavior was to warn that a strategy was not
registered, then continue creating the route without the specified auth
options. With this change, the route options will be added as long as
they are specified. If the auth strategy is not registered, then an
error will be thrown when registering the route. This seems to be inline
with README.

2. Hapi 19 no longer allows uncompiled top-level validation schemas
unless a validator is configured with the server. In previous versions,
Joi was used by default. To address this, after the validators are
created for a path/operation, any top-level object that is not already a
Joi schema is wrapped in Joi.object().

* Upgrade joi/enjoi

Upgrade enjoi to latest and joi to latest that is supported by enjoi.
Additional work needs to be done once enjoi supports @hapi/joi >= 16.

* Upgrade hoek to 9.0.4

* Upgrade swagger-parser to 9.0.1

* Upgrade boom to 9.1.0

* Upgrade nyc

This upgrade addresses a vulnerability found by npm audit

* Downgrade @hapi/joi to v14

This aligns @hapi/joi with the peer dependency requirement for enjoi.

* Update package.json to reflect node >= 12 requirement

Hapi v19 requires node >= 12. This change will make that more apparent
to consumers.

* Remove node 8 & 10 from travis

Co-authored-by: Tim Burch <tpburch@sep.com>