Branch fixtures and update tests
Each of the fixtures/defs have been branched to oas2 and oas3 versions.
Rather than using the "official" petstore OpenAPI 3 samples, I converted
the existing ones to ensure they are the same functionally.

Most of the unit tests have been updated to run both oas2 and oas3
versions to ensure all existing functionality works for both versions.

There are several issues that still need to be solved:

- oas3 servers: Right now, I am assuming there will be one server and
it's value will essentially be "basePath". I'd like to change this to
handle a full server declaration, probably using URL to parse it and get
to the correct basePath.

- oas3 file parameters: These are pretty different in oas3 and oas2. I
haven't looked to deeply at this yet - still a TODO.

- parameter serialization: In oas3, "collectionFormat" is used. Is oas3,
    this is split between "style" and "explode". Existing unit test
    cases are passing, but that is mostly a happy coincidence. Better
    support needs to be added.

- parameter schemas/templates: I'd like to find a better way to unify
oas2 and oas3 parameter structures so we can avoid a bunch of "if"
checks everywhere. I was initially trying to flatten oas3 parameters
to be more like oas2. I am now thinking that it might be better to do
the opposite and separate "parameter" metadata and the parameter
"schema".